requirements.txt
Django==5.1.6
djangorestframework==3.15.2
pandas==2.2.3
django-filter==25.1
numpy==2.2.3